Added Product Search

+++ Shoping_Card_DB/dbo/Stored Procedures/SP_CreateNewProduct.sql
@@ -2,7 +2,7 @@
 	@name nvarchar(50),
 	@amount int,
 	@info nvarchar(200),
-	@price int,
+	@price decimal,
 	@categoryId int
 as
 begin
